feat: use asset-card component in cart and preview (#507)
