+++ packages/client/hmi-client/index.html
@@ -1,6 +1,5 @@
 <!DOCTYPE html>
 <html lang="en">
-
 <head>
 	<meta charset="UTF-8" />
 	<link rel="icon" type="image/png" href="src/assets/images/icon.png" />
@@ -8,7 +7,5 @@
 	<title>TERArium</title>
 	<script type="module" src="src/main.ts"></script>
 </head>
-
 <body></body>
-
 </html>
